requirements.txt
anyio==3.6.2
appnope==0.1.3
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==22.8.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
fastapi==0.89.1
h11==0.14.0
idna==3.4
importlib-metadata==4.12.0
ipykernel==6.16.0
ipython==7.34.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.5
numpy==1.21.6
outcome==1.2.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pathspec==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
pyzmq==24.0.1
requests==2.28.1
retrying==1.3.3
selenium==4.1.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
starlette==0.22.0
tomli==2.0.1
tornado==6.2
traitlets==5.4.0
trio==0.21.0
trio-websocket==0.9.2
typed-ast==1.5.4
typing-extensions==4.3.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
uvicorn==0.20.0
wcwidth==0.2.5
wsproto==1.2.0
zipp==3.8.1
